Modularize the code
Use JS modules to separate the code in separate code. Could still use some work, since logic and dom manipulation is quite mixed in the code, but this was the best I could do with keeping the functions mostly as-is.
